plover: Escape key no longer disables auto spacing/capitalization
I'm doing an experiment where I stay in the default Plover mode all the time and
use pass-through chords for vim commands, rather than switching to "manual
mode" (automatic spacing and capitalization off) every time I press Escape.

I'm spending a lot of time typing the chord to return to the default Plover
mode, because I don't have any kind of visual mode indicator and I'm often not
confident about which of the two modes I'm in. I'm also finding that a lot of
the Plover meta-commands don't work as expected/desired in lowercase mode or
when the space character is set to the empty string (both of which are true when
I'm in "manual mode").