emacs-redux: load org-roam in org-mode hook
Use the org-mode hook to load org-roam when org-mode is started (which
happens the first time an .org file is opened or an org-related
keybinding is used).

Ideally, org-roam should be loaded even earlier, so that
org-roam--build-cache-async has a chance to finish before I start
using org-roam keybindings. (org-roam refuses to do anything until the
cache is fully built, which currently takes a few seconds.)

For the time being, I don't like the idea of loading org-roam at
startup (so I can build the cache early), because that also entails
loading org-mode at startup, which would add a couple of seconds to my
emacs startup time (boo!).