emacs-redux: remove ~/emacs.d/elpa/org-20200224 [remove]
Org-mode is now included in emacs, although I'm not sure which
emacs version was the first to include it.

When I upgraded to from Emacs 25 to Emacs 27.2, I was getting some
weird visual glitches when viewing org files (highlighting of code
blocks). When I removed the ~/.emacs.d/elpa/org-20200224 directory the
problems went away.