emacs-redux: enable evil-collection in all major modes
Enable evil-collection keybindings in all emacs major modes.

I found that enabling evil-collection keybindings for in a subset of
major modes (just vterm in my case) didn't work well in practice.

The problem was that evil-collection requires setting the following
values before loading evil and evil-collection:

  (setq evil-want-keybinding nil
        evil-want-integration t)

And those settings radically change evil keybindings in major modes
where evil-collection is not enabled (dired and magit in my case). For
example, the single letter commands in the magit status buffer stopped
working, because they were being overridden by the default vim
keybindings.

I decided the best solution was probably just to enable
evil-collection for all major modes and see how it works for me.

In spite of the awkwardness described above, I'm still keen to use
evil-collection because it greatly improves the experience of using
evil with vterm, and I expect it will improve the usability of a lot
of other major modes as well (e.g. dired).