emacs-redux: make all org headings the same size [UX]
I found that below nested heading level 7 (i.e. org-level-7 face), org
cycles back to using org-level-1 face, org-level-2 face, and so on. This
means that the font size for heading level 8 is larger than the font
size for heading level 7, which looks very weird.

So I decided it was best to just make all the font sizes for all heading
levels the same for now (i.e. `:height 1.0`).