emacs: unbind escape key in minibuffer [keybind]
When using `evil-collection-setup-minibuffer` to enable evil
keybindings in the minibuffer, the default behaviour of the Escape key
is to quit the minibuffer. This is annoying because I often do complex
editing of shell commands in the minibuffer, and accidentally hitting
the Escape key in normal mode causes me to prematurely exit the
minibuffer and lose whatever text/command I was writing.