WIP: emacs: configure lsp for Unity/C# [WIP]
It kinda works, but:

* I need to open the project in JetBrains rider first, to force Unity
to generate the *.csproj and *.sln files. This may be fixable with the
hack provided by [1], but until I get LSP fully working it is the
least of my concerns.

* I currently need to run `M-x lsp` manually after opening each C#
file. This could be fixed by adding a csharp-mode hook.

* Unity assemblies are not being correctly loaded. I think this is
because the DLL paths in the *.csproj files (e.g.  Piglet.csproj) are
Windows paths, whereas I am running emacs/omnisharp under WSL.  One
way to fix this might be to connect to the Windows EXE for omnisharp,
rather than the Linux version that is automatically installed and used
by lsp-mode.

* From what I hear, performance could be greatly improved by compiling
emacs with native JSON parsing (install libjansson-dev and ./configure
--with-json) and native compilation (emacs 28.1 with ./configure
--with-native-compilation). But I want to get everything working
properly before I worry about performance.

[1]: https://github.com/elizagamedev/unity.el