emacs: replace ivy/counsel with vertico [refactor]
Remove ivy/counsel/swiper and replace it with a basic vertico
config. This commit is the first step towards switching over to the
vertico/orderless/consult/embark family of packages.

This commit has the benefit of simplifying my emacs config, but I do
lose some functionality in the short term:

* removed live-updating ripgrep commands (e.g. counsel-project-rg)
* removed swiper
* removed counsel-projectile commands (e.g. "SPC p p" ->
counsel-projectile-switch-project)

Note: Unfortunately, I was not able to delete the ivy/counsel/swiper
packages from ~/.emacs.d/elpa because the `lispy` package depends on
those (and I still want to use `lispy`).