emacs: upgrade `magit` for emacs-28.1 [upgrade][buildfix]
After upgrading to emacs-28.1 with native-comp, the magit commit was
failing to open with the error shown below.

Reinstalling the magit package fixed it.

--- BEGIN QUOTE ---
hint: Waiting for your editor to close the file...
Waiting for Emacs...
*ERROR*: Wrong number of arguments: ((t) (fn &optional next-buffer killed-one filepos) "Honor `with-editor-server-window-alist' (which see)." (let ((server-window (save-current-buffer (set-buffer (or next-buffer (current-buffer))) (if with-editor-mode (progn (setq with-editor-previous-winconf (current-window-configuration)))) (with-editor-server-window)))) (funcall fn next-buffer killed-one filepos))), 5
error: There was a problem with the editor '/home/benv/bin/emacsclient --socket-name=/tmp/emacs1000/server'.
Please supply the message using either -m or -F option.
--- BEGIN QUOTE ---