mbsync: disable syncing of Drafts folder [bugfix]
Disable syncing of drafts folders with the gmail/fastmail server, to
work around an annoying integration bug between mu4e and mbsync. The
bug causes an steadily-growing and unbounded number of draft message
files to be created in the drafts folder while composing an email.

See [1] for a detailed discussion of the issue.

[1]: https://github.com/djcb/mu/issues/1373#issuecomment-460063934