emacs: prevent first shelldon command from running in popup [bugfix]
Prevent the first shelldon command after emacs startup from running in
a pop-up buffer, rather than in a normal window like I prefer. (The
second and subsequent shelldon commands do indeed run in a normal
window.)

This bug/annoyance happens because `display-buffer-alist` is assigned
a new value when the shelldon package is autoloaded, and this
overloads my own value of `display-buffer-alist`, as specified in
`benv/select-window-n-and-call-function` and
`benv/call-function-same-window`.