emacs: move benv/ace-link-* function to ace-link section [refactor]
Move my winum-related ace-link functions from the `use-package winum`
block to the `use-package ace-link` block.

This is a better way of organizing things and fixes a minor issue
where the `ace-link` package would not autoload when invoking the
above-mentioned ace-link/winum functions via keybinds. (To work around
that, I previously needed to manually call `ace-link`, cancel it, then
re-type the keybinding that I wanted to use in the first place.)