emacs: load window.el on startup [tweak]
Load window.el on startup, since other package
configurations (e.g. recentf) depend on it.

Note: I changed `use-package window` to `use-package emacs` here
because `window` is not a "feature" (in emacs lingo). In other words,
there is no `(provide window)` in window.el, and so `use-package
window` is not really correct. (To my surprise, it did not cause any
errors though.)