emacs: add benv/switch-to-note-in-dir{,and-focus} [new]
Add `benv/switch-to-note-in-dir`, which prompts the user to select a
note by its title using `completing-read`, then opens the note in the
neighbor window in direction DIR. If a neighbour window doesn't
already exist in direction DIR, it is created by splitting the current
window.

`benv/switch-to-note-in-dir-and-focus` does the same, but also
switches focus to the target neighbour window at the same time.