emacs: benv/select-neighbor-window -> benv/in-neighbor-window [refactor]
Replace `benv/select-neighbor-window` with `benv/in-neighbor-window`.

The original `benv/select-neighbor-window` function simply switches
focus to a neighbor window in direction DIR, creating the neighbor
window first if needed.

`benv/in-neighbor-window` is a macro that does the same thing, but
also sets up the neighbor window state such that it will behave nicely
with `quit-window`. In particular, if the neighbor window was newly
created by splitting the current window, then calling `quit-window`
will close the neighbour window and restore focus to the original
window. In order for this to work, the user-provided lisp code that is
run in the neighbour window needs to be passed as the BODY parameter
to `benv/in-neighbor-window`. That is why `benv/in-neighbor-window` is
a macro rather than a function.