tridactyl: use same .tridactylrc file on linux/windows [refactor][simplify]
IIRC, I created a symlink from tridactyl/.tridactylrc ->
windows/tridactyl/_tridactylrc because I thought that the file needed
to be named differently (with a leading underscore) on Windows.

With recent versions of tridactyl this is no longer necessary (if it
ever was?), so we can just use the same file on both Windows and
Linux.