emacs: make shelldon buffers "visible" [tweak]
Remove leading space (" ") from shelldon output buffer names, so that
they are not treated as "invisible buffers" by emacs [1].

I'm trying to fix the behaviour of "SPC TAB"
(`benv/switch-to-previous-buffer') with shelldon buffers. Whenever I
try to switch back to a shelldon buffer with "SPC TAB", it skips over
the shelldon buffer switches to some other buffer.

I thought that making the shelldon buffers non-invisible might fix the
problem, but unfortuantely it doesn't. Still, I am keeping this commit
for now, because I have no desire for shelldon buffers to be
invisible.

[1] https://www.emacswiki.org/emacs/InvisibleBuffers