emacs: fix "SPC TAB" behaviour with shelldon buffers [bugfix]
I often use "SPC TAB" (`benv/switch-to-previous-buffer') to switch
between the current and previously visited buffer.  However, whenever
I tried to use "SPC TAB" to switch back to a shelldon buffer, it would
skip over the shelldon buffer and send me somewhere else, which was
highly annoying.

This commit finally fixes that issue. The problem turned out to be
shelldon was using `display-buffer' to show the buffer rather than
`switch-to-buffer'. If understand correctly, `switch-to-buffer'
"selects" the buffer by adding a new entry to the front of the buffer
list, whereas `display-buffer' does not.