plover: prevent unwanted spaces after modifier chords [tweak]
I usually want the first chord following a modifier chord
(e.g. `M-x` to run an interactive command in emacs) to be
"attached". That is to say, I don't want a space to be automatically
inserted before the first word that is typed after a modifier chord,
because more often than not, the modifier causes a switch to a
different text input (e.g. opening a prompt, switching to a different
application window).