emacs: fix org-roam sqlite3 hanging process [bugfix]
This commit updates the org-roam code to use the new `emacsql-sqlite`
package, rather than the `emacsql-sqlite3`package.

I did this to fix a problem with `org-roam` commands hanging emacs.
When I upgraded from emacs 27.2 -> 29.1, all org-roam commands started
to hang emacs and then fail with a timeout error after ~ 30 seconds.
For future reference, and because it may turn out to be relevant,
I was seeing this issue in WSL 1 on Windows 10.

Although I could not determine the exact cause of the problem, I'm
certain it has something to do with how the `emacsql-sqlite3` package
communicates with the `sqlite3` command-line program (via a
subprocess). I suspect there has been some change in how emacs
communicates with shell subprocesses, between emacs-27.2 and
emacs-29.1, which indirectly caused this problem.

While searching for a solution, I noticed that there is a new
`emacsql-sqlite` package that replaces `emacsql-sqlite3`, and that the
latter is no longer available on MELPA. The author of the new
`emacsql-sqlite` package (Chris Wellons) had very similar problems to
mine, and he explains in detail how he solved them in his blog post:

https://nullprogram.com/blog/2014/02/06/