emacs: fix org-roam, missing function `ucs-normalize-NFD-string` [bugfix]
I was getting an error about an undefined function
`ucs-normalize-NFD-string`, when trying to create a new org-roam note.

It looks like this was caused by a compatibility issue with
emacs-29.1 (and possibly earlier emacs 28 too). The issue is discussed
in detail at:

https://github.com/org-roam/org-roam/issues/1981

I made this commit based on the pull request that is linked from the
above issue.