emacs: make eglot/flymake diagnostics work with clangd.exe [hack]
Add "advice" to `eglot-handle-notification` to translate Windows
paths/URIs to WSL paths/URIs, so that flymake diagnostics are
correctly rendered in C++ source code buffers.

Previously, flymake was not showing any annotations because it was
using the Windows-style file URI sent by `clangd.exe` to locate the
source file (and failing).