plover: add show-stroke.py (python dictionary)
This dictionary comes from the code example at
https://pypi.org/project/plover-python-dictionary/.

It defines a special prefix stroke "STR*", which will cause the next
stroke to be printed verbatim (as it would appear in Plover's "paper
tape" dialog).

This extremely handy for writing dictionary entries in JSON files, and
also for debugging Python dictionaries.