emacs: fix consult-dir by removing compat-28.1.1.1 package [bugfix]
I think that I should have deleted this package when I upgraded to
emacs 29.1. For some reason it was causing the `consult-dir` command
to fail with a "void-function compat-string-width" error, even though
I also have the `compat-29.1.4.2` package installed (which I believe
is the correct one for my current version of emacs).

If I understand correctly, the `compat` package implements elisp
functions that are only available in newer versions of emacs, in order
to help users with older emacs versions run the newest versions of
various packages (e.g. `consult-dir`).