emacs: simplify eglot code for Windows path munging [refactor]
The purpose of this code is to automatically translate Windows paths
to Linux/WSL paths (and vice versa), so that I can use `clangd.exe` as
an LSP server while running Emacs under WSL. And it works!

This commit replaces my original elisp functions that implement the
path translations with much cleaner implementations. The original
versions worked, but were incredibly messy and difficult to
understand.