plover: complete rewrite of left-fingerspelling.py [refactor]
This script is mostly working as I envisioned now, and I'm actually
using it in my day-to-day work.

I changed many of the left-hand chords mappings to ones that were more
comfortable to type, after becoming more familiar with the various
3-letter and 4-letter chords that aren't already used by other steno
chords.

I completely rewrote main part of the script and managed to
simplify/shorten the code a bit.