import: which-window.el [import][wip]
* Description

`which-window` is a global minor mode that interactively prompts the
user for a target window, whenever Emacs wants to automatically
display a buffer (e.g. a help buffer, a shell command output buffer).

* Motivation

The main motivation of this mode is to eliminate Emacs' annoying
behaviour of automatically displaying buffers (e.g. help buffers) in
unpredictable locations within your window layout. This automatic
display behaviour often messes up your carefully constructed window
arrangement, and obscures buffers that you wanted to remain visible.
As a workaround, people often use the `winner-undo` function from the
built-in `winner` package [1], to undo the unwanted changes that Emacs
makes to their window layout. This is very helpful, but not a
completely satisfying solution in my opinion. Another Emacs package
that people sometimes use to address these concerns is `shackle` [2].

In theory, the user has complete control of where Emacs' displays
buffers in any given situation, by appropriately customizing the
display rules in `display-buffer-alist`. In practice,
`display-buffer-alist` is extremely complicated and difficult to
use. This is unfortunate because window management is one of the most
obvious aspects of Emacs behaviour that one might want to customize.

As an alternative to using this mode, the user may want to try
using the built-in `display-buffer-alist` functionality to solve their
personal problems/annoyances with Emacs window management. For a
gentle introduction to customizing `display-buffer-alist`, see Prot's
excellent tutorial video at [3]. If you want to go a bit deeper, I
recommend reading the "Demystifying Emacs’s Window Manager" article from
Mickey Peterson's "Mastering Emacs'" series [4]. And finally, there
is "Displaying a Buffer in a Suitable Window" from the official
Emacs documentation [5].

* Implementation

This package is implemented by "advising" [6] Emacs' built-in
`display-buffer' function, which is invoked whenever elisp code tries
to display a buffer to the user. Instead of automatically displaying
the buffer (as per default Emacs' behaviour), immediate display of the
buffer is suppressed, and the user is instead show an interactive
prompt to select the target window for displaying the buffer.  The
interactive prompt is displayed using the `transient` package, which
is a built-in package in Emacs 29 or newer.

[1]: "Convenience Features for Window Handling" (`winner-mode`),
https://www.gnu.org/software/emacs/manual/html_node/emacs/Window-Convenience.html.
[2]: "Shackle" package for Emacs, https://depp.brause.cc/shackle/
[3]: "Emacs: control where buffers are displayed (the
'display-buffer-alist')", https://protesilaos.com/codelog/2024-02-08-emacs-window-rules-display-buffer-alist/.
[4]: "Demystifying Emacs’s Window Manager",
https://www.masteringemacs.org/article/demystifying-emacs-window-manager.
[5]: "Displaying a Buffer in a Suitable Window", https://www.gnu.org/software/emacs/manual/html_node/elisp/Displaying-Buffers.html.
[6]: "Advising Emacs Lisp Functions", https://www.gnu.org/software/emacs/manual/html_node/elisp/Advising-Functions.html.