guix: add `my-channel` with (unmodified) copy of `emacs-exwm`
I don't like the behaviour of Guix's built-in `emacs-exwm` package,
because it requires putting the `exwm` configuration in a `~/.exwm`
file, rather than in my emacs `init.el` file (like I would have
expected).

In order to customize the behaviour, I'm trying create my own modified
copy of the `emacs-exwm` package called `benv-exwm`. This commit
represents the first step, which is to make an unmodified copy of the
package and get guix to read it from my own (local) channel called
`my-channel`.

Assuming that guix can successfully install the `benv-exwm` package
from my channel, the next step is to modify the `benv-exwm` package to
have the behaviour that I actually want.