emacs: upgrade gptel (and transient)
I wanted to upgrade gptel because it's now got an improved UI and
support for media files (e.g. images).

Note: After upgrading `gptel` alone, I started to get a
transient-related error when running `gptel-menu`, as discussed in
https://github.com/karthink/gptel/issues/348:

```
transient-setup: Wrong type argument: (or eieio-object cl-structure-object oclosure), ""
```

In order to fix the above error, I needed to upgrade `transient`
package as well. However, since `transient` (version 0.4.3) is a
built-in package for Emacs 29.4, `transient` is not shown in the
completion list when running M-x `package-upgrade`. To force emacs to
upgrade a built-in package like `transient`, I needed to run
`package-upgrade` with the prefix argument.