plover: import `ben-fingerspelling.py` [import]
Custom fingerspelling dictionary with spacing control and left-hand
chords for the letters e/u/i.

Note: This dictionary is meant to be used of in conjunction with a QMK
firmware tweak I made to remap the lower-right `*` key to
`TSDZ`. `TSDZ` is the "unique ender" chord for
`ben-fingerspelling.py`, and without that key remapping, the
fingerspelling chords are not very easy to type.