emacs: upgrade `consult` to version 1.9
My `consult-buffer` command was failing with an undefined variable
error (something like "compat-string-width not defined"), so I
upgraded the `consult` package to fix it.

Another reason I want to update `consult` is that I'm interested in
trying the `consult-omni` package for doing web searches from inside
emacs, and it's probably best if I have the latest stable release of
`consult` for that.