emacs: install `consult-omni` package (for web searches)
Note: `consult-omni` is not yet available on MELPA, so I installed it
using `package-vc-install`.

I tried to rename the `~/.emacs.d/elpa/consult-omni` directory to
`~/.emacs.d/elpa/consult-omni-$COMMITHASH` (for better
record-keeping), but then`use-package` failed to load the package for
some reason. (And it's not just a matter of updating the path for the
`sources` subdirectory -- I fixed that.)