emacs: fix broken Tab-completion in minibuffer
I added a Tab-key binding to fix a problem in magit buffers, but I
made the key binding too general (by not specifying a specific
keymap), and as a result I broke the Tab key binding for shell/command
completion in the minibuffer.

I probably broke the Tab key binding in a bunch of other modes too,
but the minibuffer is place where I noticed the issue.