emacs: remove `emacsql-sqlite` binary and add to gitignore [workflow]
In order for `org-roam` to work, I need to manually compile this
binary whenever I'm setting up a new machine, by going into the
`emacs-redux/.emacs.d/elpa/emacsql-sqlite-20230225.2205/sqlite`
directory and running `make`.

I'm adding this file to `.gitignore` because I'm tired of it showing
up in my unstaged changes and getting in the way when I want to `git
pull`.